- Fixed cg_net_uri_* using size_t instead of int. - Fixed some compiler warnings using LLVM 3.3 on XCode 4.2.1.

+++ include/cybergarage/net/curi.h
@@ -132,9 +132,9 @@ BOOL cg_net_uri_isunreservedchar(char c);
 BOOL cg_net_uri_isescapechar(char c);
 BOOL cg_net_uri_isalphanumchar(char c);
 
-BOOL cg_net_uri_isescapedstring(char *buf, int bufSize);
-char *cg_net_uri_escapestring(char *buf, int bufSize, CgString *retBuf);
-char *cg_net_uri_unescapestring(char *buf, int bufSize, CgString *retBuf);
+BOOL cg_net_uri_isescapedstring(char *buf, size_t bufSize);
+char *cg_net_uri_escapestring(char *buf, size_t bufSize, CgString *retBuf);
+char *cg_net_uri_unescapestring(char *buf, size_t bufSize, CgString *retBuf);
 BOOL cg_net_uri_isequivalent(char *url, char *relative_url);
 
 CgDictionary *cg_net_uri_getquerydictionary(CgNetURI *uri);
